requirements.txt
django-rest-enumfield==0.2.0
django-rest-witchcraft==0.10.3
djangorestframework==3.12.1
jsonpath==0.82
mysqlclient==2.0.1
numpy==1.19.4
pandas==1.1.4
python-dateutil==2.8.1
python-editor==1.0.4
python-http-client==3.3.1
requests==2.25.0
sqlacodegen==2.3.0
SQLAlchemy==1.3.19
sqlparse==0.4.0
urllib3==1.26.1